# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wisp API",
    "version": "0.0.1",
    "description": "Phase 2 Fastify API for Wisp on Base Sepolia"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Health, DB, RPC, indexer lag",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/config": {
      "get": {
        "summary": "Public deployment configuration",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/openapi.json": {
      "get": {
        "summary": "OpenAPI document",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/session/sync": {
      "post": {
        "summary": "Sync profile identities and pending deliveries",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "summary": "Current profile, identities, wallet",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/wallet/challenge": {
      "post": {
        "summary": "Issue wallet link EIP-712 challenge",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/wallet/link": {
      "post": {
        "summary": "Verify and bind wallet",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/wallet/unlink": {
      "post": {
        "summary": "Soft-unlink wallet",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/resolve": {
      "post": {
        "summary": "Signed recipient descriptor",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/gifts": {
      "post": {
        "summary": "Create gift draft",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List sender gifts",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/gifts/{id}/submitted": {
      "post": {
        "summary": "Attach funding tx hash",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/gifts/{id}": {
      "get": {
        "summary": "Gift projection",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/gifts/{id}/claim-authorization": {
      "post": {
        "summary": "Recipient claim EIP-712 auth",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/gifts/{id}/read": {
      "post": {
        "summary": "Mark inbox item read",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/inbox": {
      "get": {
        "summary": "Wisp Inbox",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/activity": {
      "get": {
        "summary": "Incoming and sent activity",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Client error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}